# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.BulkGivePointRequest:
    properties:
      reasonId:
        type: string
      studentIds:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - reasonId
    - studentIds
    type: object
...
  dto.CreatePointReasonRequest:
    properties:
      name:
        type: string
      score:
        minimum: 1
        type: integer
      type:
        enum:
        - REWARD
        - PENALTY
        type: string
    required:
    - name
    - score
    - type
    type: object
...
  dto.GivePointRequest:
    properties:
      reasonId:
        type: string
      studentId:
        type: string
    required:
    - reasonId
    - studentId
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/dto.Pagination'
      success:
        type: boolean
    type: object
  dto.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  dto.PointReasonResponse:
    properties:
      id:
        type: string
      name:
        type: string
      score:
        type: integer
      type:
        type: string
    type: object
  dto.PointResponse:
    properties:
      cancelled:
        type: boolean
      cancelledAt:
        type: string
      givenAt:
        type: string
      givenBy:
        $ref: '#/definitions/dto.UserResponse'
      id:
        type: string
      reason:
        $ref: '#/definitions/dto.PointReasonResponse'
      student:
        $ref: '#/definitions/dto.StudentResponse'
    type: object
  dto.PointSummary:
    properties:
      netScore:
        type: integer
      studentId:
        type: string
      totalPenalty:
        type: integer
      totalReward:
        type: integer
    type: object
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  dto.StudentResponse:
    properties:
      createdAt:
        type: string
      grade:
        type: integer
      id:
        type: string
      name:
        type: string
      roomNumber:
        type: string
      studentNumber:
        type: string
    type: object
...
  dto.UpdatePointReasonRequest:
    properties:
      name:
        type: string
      score:
        minimum: 1
        type: integer
      type:
        enum:
        - REWARD
        - PENALTY
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 기숙사 관리 시스템 API - 학생 상벌점 관리 및 당직 일정 운영
  title: Dormi API
  version: "1.0"
paths:
...
  /point-reasons:
    get:
      description: 모든 상벌점 사유 조회
      parameters:
      - description: 유형 (REWARD, PENALTY)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PointReasonResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 상벌점 사유 목록
      tags:
      - 상벌점사유
    post:
      consumes:
      - application/json
      description: 새로운 상벌점 사유 등록
      parameters:
      - description: 사유 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePointReasonRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PointReasonResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 상벌점 사유 생성
      tags:
      - 상벌점사유
  /point-reasons/{id}:
    delete:
      description: 상벌점 사유 삭제
      parameters:
      - description: 사유 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 상벌점 사유 삭제
      tags:
      - 상벌점사유
    get:
      description: ID로 상벌점 사유 조회
      parameters:
      - description: 사유 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PointReasonResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 상벌점 사유 상세
      tags:
      - 상벌점사유
    put:
      consumes:
      - application/json
      description: 상벌점 사유 수정
      parameters:
      - description: 사유 ID
        in: path
        name: id
        required: true
        type: string
      - description: 수정할 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePointReasonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PointReasonResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 상벌점 사유 수정
      tags:
      - 상벌점사유
  /points:
    get:
      description: 상벌점 목록 조회 (필터링 지원)
      parameters:
      - description: 학생 ID
        in: query
        name: studentId
        type: string
      - description: 유형 (REWARD, PENALTY)
        in: query
        name: type
        type: string
      - description: 시작일 (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: 종료일 (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - default: 1
        description: 페이지
        in: query
        name: page
        type: integer
      - default: 20
        description: 페이지당 개수
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PointResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 상벌점 목록 조회
      tags:
      - 상벌점
    post:
      consumes:
      - application/json
      description: 학생에게 상벌점 부여
      parameters:
      - description: 상벌점 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GivePointRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PointResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 상벌점 부여
      tags:
      - 상벌점
  /points/{id}/cancel:
    patch:
      description: 상벌점 취소 (soft delete)
      parameters:
      - description: 상벌점 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 상벌점 취소
      tags:
      - 상벌점
  /points/bulk:
    post:
      consumes:
      - application/json
      description: 여러 학생에게 동시에 상벌점 부여
      parameters:
      - description: 다건 상벌점 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BulkGivePointRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PointResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 상벌점 다건 부여
      tags:
      - 상벌점
  /points/reset:
    delete:
      description: 모든 상벌점 삭제 (관리자 전용)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 상벌점 전체 초기화
      tags:
      - 상벌점
  /points/student/{studentId}:
    get:
      description: 특정 학생의 상벌점 목록 조회
      parameters:
      - description: 학생 ID
        in: path
        name: studentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PointResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 학생별 상벌점 조회
      tags:
      - 상벌점
  /points/student/{studentId}/summary:
    get:
      description: 특정 학생의 상벌점 요약 정보 조회
      parameters:
      - description: 학생 ID
        in: path
        name: studentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PointSummary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 학생별 상벌점 요약
      tags:
      - 상벌점
...
securityDefinitions:
  BearerAuth:
    description: JWT 토큰 (Bearer {token})
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"